+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.1 under development
 
+- Chg #316: Fix exception messages (@xepozz)
 - Bug #317: Fix delegated container (@xepozz)
 
 ## 1.2.0 November 05, 2022